+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Pass missing `captureFailedRequests` param to `FailedRequestInterceptor` ([#2744](https://github.com/getsentry/sentry-dart/pull/2744))
+- Bind root screen transaction to scope ([#2756](https://github.com/getsentry/sentry-dart/pull/2756))
 
 ## 8.14.0-beta.1